+++ CHANGELOG.md
@@ -1010,6 +1010,8 @@ the Nodes and Edges [epic](https://github.com/OpenFn/Lightning/issues/793).
   [#536](https://github.com/OpenFn/Lightning/issues/536)
 - Function to import a project
   [#574](https://github.com/OpenFn/Lightning/issues/574)
+- Scrub run_result dataclip on dataclip viewer
+  [#1606](https://github.com/OpenFn/Lightning/issues/1606)
 
 ### Changed